partially fix outdated detection.
if sysview receives a report from HOST A it also looks for cache files
older than a day and marks them outdated. But since we only render
the received hosts detail page and the overview the outdated status will
only be displayed on the index page until a report from the outdated host
updates the host and if there is still outdated items after that.